// src/app/models/KeHoachVonPhiBNCT.ts
import {Validators} from "@angular/forms";

export class KeHoachVonPhiBNCT {
  id: number;
  khVpBnId: number;
  maBoNganh: string;
  tenBoNganh: string;
  soDnQt: number = 0;
  soTdQt: number = 0;
  soHuyDt: number = 0;
  soTuPhaiNop: number = 0;
  soDuNamSau: number = 0;
  soTuNamSau: number = 0;
  soDtDauNam: number = 0;
  soDtBsTrongNam: number = 0;
  soDuDtNamTruoc: number = 0;
  soTuNamTruoc: number = 0;
  type: string;
}